% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENIE3.R
\name{GENIE3}
\alias{GENIE3}
\title{GENIE3}
\usage{
GENIE3(expr.matrix, tree.method = "RF", K = "sqrt", ntrees = 1000,
  regulators = NULL, ncores = 1, verbose = FALSE, seed = NULL)
}
\arguments{
\item{expr.matrix}{Expression matrix (genes x samples). Every row is a gene, every column is a sample.}

\item{tree.method}{Tree-based method used. Must be either "RF" for Random Forests (default) or "ET" for Extra-Trees.}

\item{K}{Number of candidate regulators randomly selected at each tree node (for the determination of the best split). Must be either "sqrt" for the square root of the total number of candidate regulators (default), "all" for the total number of candidate regulators, or a stricly positive integer.}

\item{ntrees}{Number of trees in an ensemble for each target gene. Default: 1000.}

\item{regulators}{Subset of genes used as candidate regulators. Must be either a vector of indices, e.g. \code{c(1,5,6,7)}, or a vector of gene names, e.g. \code{c("at_12377", "at_10912")}. The default value NULL means that all the genes are used as candidate regulators.}

\item{ncores}{Number of cores to use for parallel computing. Default: 1.}

\item{verbose}{If set to TRUE, a feedback on the progress of the calculations is given. Default: FALSE.}

\item{seed}{Random number generator seed for replication of analyses. The default value NULL means that the seed is not reset.}
}
\value{
Weighted adjacency matrix of inferred network. Element w_ij (row i, column j) gives the importance of the link from regulatory gene i to target gene j.
}
\description{
\code{GENIE3} Infers a gene regulatory network (in the form of a weighted adjacency matrix) from expression data, using ensembles of regression trees.
}
\examples{
## Generate fake expression matrix
exprMatrix <- matrix(sample(1:10, 100, replace=TRUE), nrow=20)
rownames(exprMatrix) <- paste("Gene", 1:20, sep="")
colnames(exprMatrix) <- paste("Sample", 1:5, sep="")

## Run GENIE3
weightMatrix <- GENIE3(exprMatrix, regulators=paste("Gene", 1:5, sep=""))

## Get ranking of edges
linkList <- get.link.list(weightMatrix)
head(linkList)
}

