% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.link.list.R
\name{get.link.list}
\alias{get.link.list}
\title{get.link.list}
\usage{
get.link.list(weight.matrix, report.max = NULL, threshold = 0)
}
\arguments{
\item{weight.matrix}{Weighted adjacency matrix as returned by \code{\link{GENIE3}} or \code{\link{tlGENIE3}}.}

\item{report.max}{Maximum number of links to report. The default value NULL means that all the links are reported.}

\item{threshold}{Only links with a weight above the threshold are reported. Default: threshold = 0, i.e. all the links are reported.}
}
\value{
List of regulatory links in a data frame. Each line of the data frame corresponds to a link. The first column is the regulatory gene, the second column is the target gene, and the third column is the weight of the link.
}
\description{
\code{get.link.list} Converts the weight matrix, as returned by \code{\link{GENIE3}} or \code{\link{tlGENIE3}}, to a sorted list of regulatory links (most likely links first).
}
\examples{
## Generate fake expression matrix
expr.matrix <- matrix(sample(1:10, 100, replace=TRUE), nrow=20)
rownames(expr.matrix) <- paste("Gene", 1:20, sep="")
colnames(expr.matrix) <- paste("Sample", 1:5, sep="")

## Run GENIE3
weight.matrix <- GENIE3(expr.matrix, regulators=paste("Gene", 1:5, sep=""))

## Get ranking of edges 
link.list <- get.link.list(weight.matrix)
head(link.list)
}
\seealso{
\code{\link{GENIE3}}, \code{\link{tlGENIE3}}
}

