% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinkList.R
\name{getLinkList}
\alias{getLinkList}
\title{getLinkList}
\usage{
getLinkList(weightMatrix, reportMax = NULL, threshold = 0)
}
\arguments{
\item{weightMatrix}{Weighted adjacency matrix as returned by \code{\link{GENIE3}}.}

\item{reportMax}{Maximum number of links to report. The default value NULL means that all the links are reported.}

\item{threshold}{Only links with a weight above the threshold are reported. Default: threshold = 0, i.e. all the links are reported.}
}
\value{
List of regulatory links in a data frame. Each line of the data frame corresponds to a link. The first column is the regulatory gene, the second column is the target gene, and the third column is the weight of the link.
}
\description{
\code{getLinkList} Converts the weight matrix, as returned by \code{\link{GENIE3}}, to a sorted list of regulatory links (most likely links first).
}
\examples{
## Generate fake expression matrix
exprMat <- matrix(sample(1:10, 100, replace=TRUE), nrow=20)
rownames(exprMat) <- paste("Gene", 1:20, sep="")
colnames(exprMat) <- paste("Sample", 1:5, sep="")

## Run GENIE3
weightMat <- GENIE3(exprMat, regulators=paste("Gene", 1:5, sep=""))

## Get ranking of edges
linkList <- getLinkList(weightMat)
head(linkList)
}
\seealso{
\code{\link{GENIE3}}
}
