% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlGENIE3.R
\name{tlGENIE3}
\alias{tlGENIE3}
\title{tlGENIE3}
\usage{
tlGENIE3(TS.data, SS.data = NULL, tree.method = "RF", K = "sqrt",
  ntrees = 1000, h = 1, regulators = NULL, ncores = 1,
  verbose = FALSE, seed = NULL)
}
\arguments{
\item{TS.data}{List of expression matrices, where each matrix (genes x time points) corresponds to a time series experiment. Each row of a matrix is a gene, each column is a time point. The number of time points does not need to be the same in all the experiments, but the genes must be the same.}

\item{SS.data}{Steady-state expression matrix (genes x samples). Every row is a gene, every column is a sample. The default value NULL means that no steady-state data are used for network inference.}

\item{tree.method}{Tree-based method used. Must be either "RF" for Random Forests (default) or "ET" for Extra-Trees.}

\item{K}{Number of candidate regulators randomly selected at each tree node (for the determination of the best split). Must be either "sqrt" for the square root of the total number of candidate regulators (default), "all" for the total number of candidate regulators, or a stricly positive integer.}

\item{ntrees}{Number of trees in an ensemble for each target gene. Default: 1000.}

\item{h}{Value of the time lag (i.e. number of time points between the expressions of the candidate regulators and the expression of the target gene). Default: 1.}

\item{regulators}{Subset of genes used as candidate regulators. Must be either a vector of indices, e.g. \code{c(1,5,6,7)}, or a vector of gene names, e.g. \code{c("at_12377", "at_10912")}. The default value NULL means that all the genes are used as candidate regulators.}

\item{ncores}{Number of cores to use for parallel computing. Default: 1.}

\item{verbose}{If set to TRUE, a feedback on the progress of the calculations is given. Default: FALSE.}

\item{seed}{Random number generator seed for replication of analyses. The default value NULL means that the seed is not reset.}
}
\value{
Weighted adjacency matrix of inferred network. Element w_ij (row i, column j) gives the importance of the link from regulatory gene i to target gene j.
}
\description{
\code{tlGENIE3} Infers a gene regulatory network (in the form of a weighted adjacency matrix) from time series of expression data, using ensembles of regression trees. The method can also be used to infer a network jointly from timeseries and steady-state expression data.
}
\examples{
## Generate fake time series of expression data
data1 <- matrix(sample(1:10, 70, replace=TRUE), nrow=10)
rownames(data1) <- paste("Gene", 1:10, sep="")
colnames(data1) <- paste("Time point", 1:7, sep="")

data2 <- matrix(sample(1:10, 50, replace=TRUE), nrow=10)
rownames(data2) <- paste("Gene", 1:10, sep="")
colnames(data2) <- paste("Time point", 1:5, sep="")

TS.data <- list(data1,data2) 

## Run tlGENIE3
weight.matrix <- tlGENIE3(TS.data)

## Get ranking of edges 
link.list <- get.link.list(weight.matrix)
head(link.list)
}

